Fixed bug#22680839: DEFAULT IS NOT DETERMINISTIC AND SHOULD NOT BE ALLOWED IN GENERATED COLUMNS
Since it is possible to change the DEFAULT value for a column using ALTER
TABLE, the DEFAULT function may not be considered deterministic.  This can
cause inconsistent results if the DEFAULT function is used in definition of
generated columns after changing the DEFAULT value for a column.  To avoid
this problem, this patch make table rebuild if any column changes the default
value and there is a dependency on DEFAULT function in a generated expression.