Bug #19641806 	SQL_AUTO_IS_NULL IS NOT REPLICATED, FAILURE IN RPL_SESSION_VAR AND RPL_INSERT_ID
The reason of the tests failing was an incorrect initialization of
THD::substitute_null_with_insert_id member by the binlog or the slave applier
threads.
That had an effect of the very first group of events including
a Query-log-event dependent on LAST_INSERT_ID to be replayed
by the applier incorrectly as LAST_INSERT_ID was simply
ignored.
The such way chosen initialization was used to work correctly in
combination with another piece of bug14553 patch, commit:2c9f5cc706, which was
refactored/removed by wl3146.
WL3146 merge with bug14553 fixes made some efforts,
commit:799ecc9f2abeb19db7d2409ac756f627b25b75a3, to correct
behavior to add up the member set to TRUE
in THD::cleanup_after_query() when the current statement has turned
THD::first_successful_insert_id_in_cur_stmt to a non-default value.

Notice that in the applier cases this method can't work
for a reason that the applier session can break at any time to leave the effects
of the autoincrement INSERT gone, therefore invisible to a following session
like in the following example:

--connection applier_0
  create table t1(a int auto_increment primary key);
  INSERT into t1 set a=null
--disconnect

--connection applier_1
  set @@session.sql_auto_is_null=1;
  set @@session.last_insert_id=1;
  INSERT into t2 SELECT a from t1 where a is null;

The bug would manifest now as

  SELECT * from t2 => \empty

It applies to both the binlog applier and the slave applier.
The difference is in the way set @@session.last_insert_id is represented.

**Fixed** to flagging THD::substitute_null_with_insert_id up
along with setting THD::first_successful_insert_id_in_prev_stmt to a non-default
value which happens at handling of set @@session.last_insert_id and the corresponding
Intvar event.

Kuudos to LiBing who spotted the right reason of this tricky
artifact and Guilhem for correcting the initial patch!