Bug#22299443 OBSOLETE INDEX DATA SENT TO GETNDBINDEXOPERATION() CRASHES NDBD
When a transaction with a NdbIndexOperation based on an obsolete unique
index is executed, it causes a segfault in ndbd. This is due to the
access of the unseized or already dropped index from c_theIndexes
ArrayPool in Dbtc::readIndexTable(). The index was not being validated
by the readIndexTable function.

The fix is to check if the index is in use or not during the execution
of the transaction. Since the indexState enum doesn't have a separate
value to mark the not in use indexes, a new value IS_OFFLINE is now
added to the enum.

During the initialisation of Dbtc, all the indexes are seized, marked as
IS_OFFLINE and then released. Also when an index is dropped, it is
marked as IS_OFFLINE.

So now when an obsolete index is passed to obtain an NdbIndexOperation,
its state is checked and the transaction goes through only if it is not
IS_OFFLINE. If the index state is found to be IS_OFFLINE, the resources
are freed and a proper error is sent back.