BUG#22915003 - JOINING TERMINATING THREADS SHOULD NOT BE DONE IN NEW THREAD CREATION.
Description:
In the current implementation, the terminating worker threads are being joined in a function that get's called also when new threads are being created.
This function blocks on a list mutex, which can cause a bottleneck for threads (connections) creation.

Solution:
Removed joining the threads from create_thread and  shorten the timeout for sweeping timer.

RB: 12092
Reviewed by: Jan Kneschke <jan.kneschke@oracle.com>
Reviewed by: Alfredo Kojima <alfredo.kengi.kojima@oracle.com>