Bug #22679185   INVALID INNODB FTS DOC ID DURING INSERT
Problem:
=======
Insert into an InnoDB table that contains a FULLTEXT KEY and a
FTS_DOC_ID column fail when the specified FTS_DOC_ID value as high.
Currently the behaviour is consecutive doc_ids difference in the
table should not exceed 10000.

Solution:
========
Increased the limit for consecutive doc_ids difference to 65535.

Reviewed-by: Jimmy Yang <jimmy.yang@oracle.com>
RB: 11997