Bug#22810883: ASSERTION FAILED: !(USED_TABS & (~READ_TABLES & ~FILTER_FOR_TABLE))
If the generation expression of an indexed generated column contains
an AND or OR expression, the optimizer may choose that index too often
and create execution plans that give wrong results.

The problem is that Item_cond::eq() does not inspect the arguments of
the AND or OR expression. It inspects the "args" member, but it is
always empty for Item_cond objects. They instead keep their arguments
in the member "list".

The result is that it always considers two AND expressions as equal.
Similarly, it always considers two OR expressions as equal. Because of
this, the optimizer may choose to replace an AND or OR expression, for
example in a WHERE clause, with a generated column that is not
equivalent to the replaced expression.

The fix is to change Item_cond::eq() to inspect the "list" member
instead of the "args" member.