WL#7677: Improve event buffer memory allocation
The existing event buffer allocation algorith hardly ever
deallocate any memory, but keep it in its own free-list
for later reuse.

This result in the total memory usage accumulating to
the total max ever used.

This WL fix that by allowing the event memory buffers to
be returned to the operating system when the events has
been consumed.