wl#7677, Yet another followup patch... sigh
This time the usage of MAP_ANONYMOUS caused build break on
osx and Windows, which seems to not support this. On windows
as my_mmap() is not natively supported by Windows, and we
use a my_mmap() function instead. That does not implement
MAP_ANONYMOUS functionality. (Possibly could be implemented
later, but that is out of scope for this patch)

On osx it is ,,,, well not supported.

non-ANONYMOUS mapping would have required is to create
a file to map agains. That does not seem like a favourable
solution for allocating this kind of buffer memory.

Instead this fix let platforms not supporting ANONYMOUS mapping
use plain malloc instead. Should not really matter much as
long as Linux (and Solaris) use mmap().

Also fixes a few compiler warnings seen on Windows.