BUG#22957191: INVALID THREAD_ID TYPE ON TRANSACTION_CONTEXT_EVENT
thread_id type on Transaction_context_event was changed from
  long long int thread_id
to
  unsigned int thread_id
on BUG 11752665: WINDOWS SERVER X64: SO MANY COMPILER WARNINGS to
fix Windows compiler warnings, though this change did break Group
Replication on Solaris Sparc platform, since the on-the-wire data
is 64 bits size and it will overflow when reading from the network
to the event.

Changed thread_id type to unsigned 64 bits integer to solve the
Solaris Sparc platform issues.