Bug#22602324	HANDLE_FATAL_SIGNAL (SIG=11) IN TRACE::FUNCTION_ENTER
Problem: When semi sync on master is not enabled, resetting
rpl_semi_sync_master_wait_no_slave variable to OFF from ON is causing
server crash.

Analysis: active_tranxs_ is a member in ReplSemiSyncMaster to maintain
active transaction list which is initiated in enableMaster() and destroyed
in disableMaster(). Many of the semi sync replication functions look at
rpl_semi_sync_master_status and proceed further to access/modify active_tranxs_ list
only if the status is ON. This implies a restriction that rpl_semi_sync_master_status
should be set ON only if active_tranxs_ is initiated (i.e., only after
rpl_semi_sync_master_enabled is ON).

As part of the bug#17510411 patch, when rpl_semi_sync_master_wait_no_slave
global variable is set to ON, we are forcefully switching ON
rpl_semi_sync_master_status without checking whether semi sync is enabled
or not. This will lead to problems when the server is trying to access active_tranxs_
(which is not yet initialized).

Fix: Before forcefully switching ON rpl_semi_sync_master_status,
server will check now whether semi sync is enabled or not. If it is
enabled, only then it will switch on rpl_semi_sync_master_status to ON.