WL#11065: Backport JSON functions from 8.0 to 5.7 (part 4)
The following bug fix is backported to 5.7 from trunk:

Bug#22961128: WRONG RESULT WHEN MANIPULATING RESULT OF JSON_MERGE

The DOM structure returned by JSON_MERGE could sometimes have
inconsistencies which confused functions such as JSON_SET and made
them return wrong results.

The problem was that Json_object::consume() didn't always set the
parent pointer of the merged nodes, so that some of the merged nodes
appeared to be root nodes even though they were not.

The fix is to make Json_object::consume() set the parent pointer of
the merged node when it has merged two object members.

Change-Id: I8fcfd53e32cfce73d429e3dc7012acb6c78a604f