Fixed bug#22241015: REPLICATION SERVERS CRASHES WHILE CHANGING A VALUE OF INDEXED JSON FIELD
When updating virtual generated columns that are BLOBs, some storage engines
require that we have both the old and new BLOB value for virtual generated
columns that are indexed in order for the storage engine to be able to
maintain the index.

During an UPDATE operation, before writing the new value of the BLOB into the
Field_blob object, we need to ensure that the old value does not get
overwritten since this is not stored in neither the record buffer or by the
storage engine. For doing this, the Field_blob::keep_old_value() function is
used. For normal record operations, this function is called from
update_generated_write_fields() (in table.cc).

This bug happens on the slave during replication of an update operation that
contains a virtual blob column that is indexed. The slave receives the new
values for all changed fields, also including the computed values for
generated fields. These values are applied to the record buffer for the
updated record. The function update_generated_write_fields() is not used for
computing the values for the generated column. Instead the values received
from replication is used. Since update_generated_write_fields() is not called,
the old blob value stored in the Field_blob object is overwritten when
storing the new value for the blob. This causes a crash when InnoDB tries
to maintain the index based on the content of the blob.  When take care
of indexed virtual generated columns in slave, we still need the same
method to update. Therefore, this patch calls the Field_blob::keep_old_value
when unpacking the record.