Bug#22912800 ST_ASGEOJSON() DOESN'T WORK WITH SESSION VARIABLES
Problem
=======
When trying to use a user-defined variable as the geometry argument
in ST_AsGeoJSON, the server would report the error "Incorrect type
for argument geojson in function st_asgeojson.". This has two flaws:

1) The argument is not a GeoJSON, but a geometry. Hence, the error
   message should state "Incorrect type for argument geometry ...".

2) The function should allow for the usage of user-defined variable
   as the argument. ST_GeoHash for instance allows this.

Fix
===
Since the ST_GeoHash already handles this situation correctly,
we move the code for checking the argument type from the GeoHash code
into a separate function. We use this function both in the GeoHash
function and in the GeoJSON function for checking the argument type.
This also ensures consistent handling of type checking the geometry
arguments in both functions.

As a part of this patch, we move all GIS functions that previously were
in item_func* and item_strfunc* over to item_geofunc*.

We also change the error message to print out "Incorrect type for
argument geometry ..." instead of "Incorrect type for argument
geojson...".