Bug #22748612 MYSQL-57 AND MYSQL-TRUNK(SLAVE) FAILS TO CONNECT TO MYSQL-55(MASTER)
Bug21455603 fixes made the new server unable to connect to
the @@global.server_uuid unaware master. The reason appears to be that the
former bugfixes left a flaw in attempt to refine a pattern for
tolerating an unexisted system variables at handshake with the master
server. The pattern dealt with SERVER_UUID and SERVER_ID as well.

Fixed with refining checks done after a failing mysql_real_query() in two
places: one for SERVER_UUID and the other for SERVER_ID.