Rename ut_crc32* function to ut_crc32c*
Current ut_crc32 function represent the CRC32C algorithm so the
rename reflects this.

This change is in preparation for addition of CRC32 functions.

The slice8 table is also abstracted in preparation for a CRC32
table.

Software implementation also extended to allow for input
CRC32 value to which the function will add to.