Fixed bug#22392268: GCOLS: CAN'T SHOW CREATE TABLE, OR CRASH IN ITEM_FUNC::WALK
During refix_gc_items, it uses the Query_arena of the whole query to fix the
generated expression. For the case on the bug report, the Item_field 'a' can
be converted into a constant value. During such a conversion, one ref object
will be created. However, at the end of the query, this object will be freed
because THD::cleanup_after_query() will free all of the Item objects in
thd->mem_root. Therefore, the newly-created ref object is freed too. After
this, the other SHOW query comes. It will use the cached TABLE objects
directly. So when fix fields, the invalid ref object will be met and cause
a problem.

The solution in this patch is to assign a seperated Query_arena during
refix_gc_items just the same as unpack_gcol_info_from_frm. By this way, we can
keep the ref object the same life span as the whole TABLE existing.

In this patch, it doesn't fix the second problem(ERROR 1267 (HY000)) on the
bug report. This is because there is incosistent character set between
constant string with default TABLE creation. I would like to leave this
problem to character set fixing. The relative solution now is that you have to
set the unified character set explicitly.