Bug#21441297  INVALIDATTRINFO RETURNED FROM EXECTUX_BOUND_INFO WHILE SCANNING
When we perform "analyze" command in a table with index, ndbmtd crashes
and throws "InvalidAttrInfo" error due to corruption in the signal.

The problem was there in the Trix block.Here we use stack variable in
building up a signal, hence it goes out of scope when the signal is
refereed in other modules. As the memory is invalid it gives a crash
when signal is de-refereed in other modules.

Code was modified such that the signal is now build with pooled
variable.As this is a global variable too it can be refereed in other
modules.