Fixed bug#22561845: gcols: read uncommitted, crash in row_sel_sec_rec_is_for_clust_rec
In function evaluate_join_record, we miss to check whether there is a status
error after check condition related with tables if the condition doesn't
match. When innodb calls my_eval_gcolumn_expr_helper to calculate the value of
generated expression, the status error might results in the function returning
error. This error results in such a bug.

In this patch, an error check is added to evaluate_join_record(). When the
condition doesn't match and there is a status error, evaluate_join_record will
return an error nestloop state(NESTED_LOOP_ERROR) directly instead of moving
forward.