Bug#22897921 TDE: BLOCKING ONLINE DDL(INPLACE) ,WHICH OTHERWISE ALLOWED IN 5.7 Bug#22939581 SERVER CRASHED IN MUTEX_CREATE DUE TO INVALID OPTION Bug#22723797 WITH TDE, MYSQLD FAILS TO START READONLY: !SRV_READ_ONLY_MODE || M_IMPL.M_LOG_M
The root cause of this bug is: we used the encrypt_type in create_info
to check the encryption attribute change.
In fact, we need to compare the old value and the new value of it.
And, in this patch we also fixed bug#22939581. It's caused by we didn't
check the return error of innobase_start_or_create_for_mysql.
We fixed bug#22723797 in this patch too, since it's a minor change. The
cause of this bug is we need to skip calling fil_encryption_rotate when
server startup in readonly mode.

Reviewed-by: Jimmy Yang<jimmy.yang@oracle.com>
RB: 12144