Bug#22971530: memcache tests are getting skipped on Daily-Trunk, Linux Gcov
Problem:The entire memcache test cases are getting skipped on Daily-trunk,linux-gcov platform
due to the reason of, the build is not memcache enabled.

solution: In order to make the build memcache enabled, one should add the parameter -DWITH_INNODB_MEMCACHED=ON
in the compile-pentium-gcov file, under BUILD folder of source tree.

Reviewed-by: Jimmy Yang <jimmy.yang@oracle.com>

RB: 12229