Bug#22579581 NDB_PRINT_FILE CORE DUMP ON SOLARIS
An pointer to a 32 bit word was accessed via time_t pointer.

On sparc-v9 this cause bus error since time_t was 64bit and word was not
64bit aligned for File_formats::Zero_page_header::m_time.

On other platforms there 64bit time_t is used, the high word of the time_t
could be non zero resulting in bogus timestamp.